Get rid of screenjournal/movie.go
Merge everything into metadata.go
